refactor: conventions plugins
