+++ build-logic/src/main/kotlin/android-core-convention.gradle.kts
@@ -1,6 +1,3 @@
-import gradle.kotlin.dsl.accessors._8e8a6dd48b2094ffcd3758431423791d.implementation
-import java.util.Properties
-
 plugins {
     id("com.android.library")
     id("org.jetbrains.kotlin.android")
@@ -15,17 +12,9 @@ android {
     composeOptions {
         kotlinCompilerExtensionVersion = AndroidConst.COMPOSE_COMPILER_VERSION
     }
-    defaultConfig {
-        val properties = Properties()
-        properties.load(File("secrets.properties").inputStream())
-        manifestPlaceholders["YANDEX_CLIENT_ID"] = properties.getProperty("YANDEX_CLIENT_ID")
-    }
 }
 
 dependencies {
-    // Yandex login sdk
-    implementation(libs.authsdk)
-
     // Dagger 2
     implementation(libs.dagger)
     ksp(libs.dagger.compiler)
@@ -56,21 +45,6 @@ dependencies {
 
     // Serialization
     implementation(libs.kotlinx.serialization.json)
-
-    // Splash screen
-    implementation(libs.androidx.core.splashscreen)
-
-    // Date picker
-    implementation(libs.material3)
-
-    // Compose navigation
-    implementation(libs.androidx.navigation.compose)
-
-    // System bar colors
-    implementation(libs.accompanist.systemuicontroller)
-
-    // Collapsing Toolbar Layout
-    implementation(libs.toolbar.compose)
 
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.lifecycle.runtime.ktx)
