--- build-logic/src/main/kotlin/android-core-ui-convention.gradle.kts
@@ -1,75 +1,0 @@
-plugins {
-    id("com.android.library")
-    id("org.jetbrains.kotlin.android")
-    id("com.google.devtools.ksp")
-}
-
-android {
-    baseAndroidConfig()
-    buildFeatures {
-        compose = true
-    }
-    composeOptions {
-        kotlinCompilerExtensionVersion = AndroidConst.COMPOSE_COMPILER_VERSION
-    }
-}
-
-dependencies {
-    // Dagger 2
-    implementation(libs.dagger)
-    ksp(libs.dagger.compiler)
-
-    // Room
-    implementation(libs.androidx.room.runtime)
-    annotationProcessor(libs.androidx.room.compiler)
-    ksp(libs.androidx.room.compiler)
-
-    // Work manager
-    implementation(libs.androidx.work.runtime.ktx)
-
-    // DataStore
-    implementation(libs.androidx.datastore.preferences)
-
-    // Ktor
-    implementation(libs.ktor.client.okhttp)
-    implementation(libs.ktor.client.logging.jvm)
-    implementation(libs.ktor.client.serialization)
-    implementation(libs.ktor.client.content.negotiation)
-    implementation(libs.ktor.serialization.kotlinx.json)
-    implementation(libs.ktor.client.auth)
-    implementation(libs.ktor.client.websockets)
-    runtimeOnly(libs.ktor.client.auth)
-    implementation(libs.ktor.client.android)
-    implementation(libs.kotlin.stdlib.jdk8)
-
-    // Serialization
-    implementation(libs.kotlinx.serialization.json)
-
-    // Splash screen
-    implementation(libs.androidx.core.splashscreen)
-
-    // Date picker
-    implementation(libs.material3)
-
-    // Compose navigation
-    implementation(libs.androidx.navigation.compose)
-
-    // System bar colors
-    implementation(libs.accompanist.systemuicontroller)
-
-    // Collapsing Toolbar Layout
-    implementation(libs.toolbar.compose)
-
-    implementation(libs.androidx.core.ktx)
-    implementation(libs.androidx.lifecycle.runtime.ktx)
-    implementation(libs.androidx.activity.compose)
-    implementation(platform(libs.androidx.compose.bom))
-    implementation(libs.androidx.ui)
-    implementation(libs.androidx.ui.graphics)
-    implementation(libs.androidx.ui.tooling.preview)
-    implementation(libs.androidx.material3)
-    androidTestImplementation(libs.androidx.junit)
-    androidTestImplementation(libs.androidx.espresso.core)
-    androidTestImplementation(platform(libs.androidx.compose.bom))
-    androidTestImplementation(libs.androidx.ui.test.junit4)
-}
